fixed the state update and refactored
